handle pseudo waiting kernels when switching kernels

+++ jupyter_server/services/sessions/sessionmanager.py
@@ -5,7 +5,10 @@
 import os
 import pathlib
 import uuid
+from asyncio import Task
 from typing import Any, Dict, List, NewType, Optional, Union, cast
+
+from requests import session
 
 KernelName = NewType("KernelName", str)
 ModelName = NewType("ModelName", str)
@@ -16,6 +19,7 @@ except ImportError:
     # fallback on pysqlite2 if Python was build without sqlite
     from pysqlite2 import dbapi2 as sqlite3  # type:ignore[no-redef]
 
+import asyncio
 from dataclasses import dataclass, fields
 
 from jupyter_core.utils import ensure_async
@@ -210,6 +214,8 @@ class SessionManager(LoggingConfigurable):
     _connection = None
     _columns = {"session_id", "path", "name", "type", "kernel_id"}
 
+    fut_kernel_id_dict: Optional[Dict[str, Task[str]]] = None
+
     @property
     def cursor(self):
         """Start a cursor and create a database called 'session'"""
@@ -267,6 +273,7 @@ class SessionManager(LoggingConfigurable):
         type: Optional[str] = None,
         kernel_name: Optional[KernelName] = None,
         kernel_id: Optional[str] = None,
+        session_id: Optional[str] = None,
     ) -> Dict[str, Any]:
         """Creates a session and returns its model
 
@@ -276,7 +283,13 @@ class SessionManager(LoggingConfigurable):
             Usually the model name, like the filename associated with current
             kernel.
         """
-        session_id = self.new_session_id()
+
+        if session_id is not None and self.fut_kernel_id_dict is None:
+            self.fut_kernel_id_dict = {}
+
+        if session_id is None or session_id == "":
+            session_id = self.new_session_id()
+        
         record = KernelSessionRecord(session_id=session_id)
         self._pending_sessions.update(record)
         if kernel_id is not None and kernel_id in self.kernel_manager:
@@ -312,6 +325,20 @@ class SessionManager(LoggingConfigurable):
         assert isinstance(path, str)
         return {**os.environ, "JPY_SESSION_NAME": path}
 
+    async def start_kernel_async(
+        self, 
+        path: Optional[str], 
+        kernel_name: Optional[KernelName],
+    ):
+        kernel_path = await ensure_async(self.contents_manager.get_kernel_path(path=path))
+        kernel_env = self.get_kernel_env(path)
+        self.log.info(f"starting kernel ${path} ${kernel_name} ${kernel_path} ${kernel_env}")
+        self.fut_kernel_id_dict[session_id] = asyncio.create_task(self.kernel_manager.start_kernel(
+            path=kernel_path,
+            kernel_name=kernel_name,
+            env=kernel_env,
+        ))
+
     async def start_kernel_for_session(
         self,
         session_id: str,
@@ -337,15 +364,61 @@ class SessionManager(LoggingConfigurable):
             the name of the kernel specification to use.  The default kernel name will be used if not provided.
         """
         # allow contents manager to specify kernels cwd
-        kernel_path = await ensure_async(self.contents_manager.get_kernel_path(path=path))
+        if self.fut_kernel_id_dict is not None:
+            if session_id in self.fut_kernel_id_dict:
+                fut_kernel_id = self.fut_kernel_id_dict[session_id]
+                if fut_kernel_id.done():
+                    try:
+                        kernel_id = await fut_kernel_id
+                        self.fut_kernel_id_dict.pop(session_id)
+                        return kernel_id
+                    except:
+                        self.log.error(f"kernel start failed, retying ${path} ${kernel_name}")
+                        await start_kernel_async()
+            else:
+                await start_kernel_async()
+            kernel_id = "waiting"
+        else:
+            kernel_path = await ensure_async(self.contents_manager.get_kernel_path(path=path))
 
-        kernel_env = self.get_kernel_env(path, name)
-        kernel_id = await self.kernel_manager.start_kernel(
-            path=kernel_path,
-            kernel_name=kernel_name,
-            env=kernel_env,
-        )
+            kernel_env = self.get_kernel_env(path, name)
+            kernel_id = await self.kernel_manager.start_kernel(
+                path=kernel_path,
+                kernel_name=kernel_name,
+                env=kernel_env,
+            )
         return cast(str, kernel_id)
+    
+    def waiting_kernel(
+        self
+    ) -> Dict[str, Any]:
+        return {
+            "id": "waiting",
+            "name": "Waiting for kernel to start",
+            "last_activity": "2024-10-02T16:56:56.423328Z",
+            "execution_state": "waiting",
+            "connections": 0,
+        }
+    
+    def waiting_session(
+        self, 
+        session_id: string,
+        path: string,
+        type: string,
+        name: string,
+    ) -> Dict[str, Any]:
+        result = {
+            "id": session_id,
+            "kernel": waiting_kernel(),
+            "path": path,
+            "type": type,
+            "name": name,
+            "notebook": {"path": path, "name": name},
+            "last_activity": "2024-10-02T16:56:56.423328Z",
+            "execution_state": "waiting",
+            "connections": 0
+        }
+        return result
 
     async def save_session(self, session_id, path=None, name=None, type=None, kernel_id=None):
         """Saves the items for the session with the given session_id
@@ -372,11 +445,14 @@ class SessionManager(LoggingConfigurable):
         model : dict
             a dictionary of the session model
         """
-        self.cursor.execute(
-            "INSERT INTO session VALUES (?,?,?,?,?)",
-            (session_id, path, name, type, kernel_id),
-        )
-        result = await self.get_session(session_id=session_id)
+        if kernel_id != "waiting":
+            self.cursor.execute(
+                "INSERT INTO session VALUES (?,?,?,?,?)",
+                (session_id, path, name, type, kernel_id),
+            )
+            result = await self.get_session(session_id=session_id)
+        else:
+            result = waiting_session(session_id, path, type, name)
         return result
 
     async def get_session(self, **kwargs):
@@ -397,37 +473,42 @@ class SessionManager(LoggingConfigurable):
             returns a dictionary that includes all the information from the
             session described by the kwarg.
         """
-        if not kwargs:
-            msg = "must specify a column to query"
-            raise TypeError(msg)
-
-        conditions = []
-        for column in kwargs:
-            if column not in self._columns:
-                msg = f"No such column: {column}"
+        session_id = kwargs["session_id"]
+        if self.fut_kernel_id_dict is not None and session_id in self.fut_kernel_id_dict:
+            model = waiting_session(session_id, "unknown", "notebook", "Waiting for kernel to start")
+        else:
+            if not kwargs:
+                msg = "must specify a column to query"
                 raise TypeError(msg)
-            conditions.append("%s=?" % column)
-
-        query = "SELECT * FROM session WHERE %s" % (" AND ".join(conditions))  # noqa: S608
-
-        self.cursor.execute(query, list(kwargs.values()))
-        try:
-            row = self.cursor.fetchone()
-        except KeyError:
-            # The kernel is missing, so the session just got deleted.
-            row = None
-
-        if row is None:
-            q = []
-            for key, value in kwargs.items():
-                q.append(f"{key}={value!r}")
-
-            raise web.HTTPError(404, "Session not found: %s" % (", ".join(q)))
-
-        try:
-            model = await self.row_to_model(row)
-        except KeyError as e:
-            raise web.HTTPError(404, "Session not found: %s" % str(e)) from e
+    
+            conditions = []
+            for column in kwargs:
+                if column not in self._columns:
+                    msg = f"No such column: {column}"
+                    raise TypeError(msg)
+                conditions.append("%s=?" % column)
+    
+            query = "SELECT * FROM session WHERE %s" % (" AND ".join(conditions))  # noqa: S608
+    
+            self.cursor.execute(query, list(kwargs.values()))
+            try:
+                row = self.cursor.fetchone()
+            except KeyError:
+                # The kernel is missing, so the session just got deleted.
+                row = None
+    
+            if row is None:
+                q = []
+                for key, value in kwargs.items():
+                    q.append(f"{key}={value!r}")
+    
+                raise web.HTTPError(404, "Session not found: %s" % (", ".join(q)))
+    
+            try:
+                model = await self.row_to_model(row)
+            except KeyError as e:
+                raise web.HTTPError(404, "Session not found: %s" % str(e)) from e
+    
         return model
 
     async def update_session(self, session_id, **kwargs):
@@ -463,8 +544,11 @@ class SessionManager(LoggingConfigurable):
             self.cursor.execute(
                 "SELECT path, name, kernel_id FROM session WHERE session_id=?", [session_id]
             )
-            path, name, kernel_id = self.cursor.fetchone()
-            self.kernel_manager.update_env(kernel_id=kernel_id, env=self.get_kernel_env(path, name))
+            try:
+                path, name, kernel_id = self.cursor.fetchone()
+                self.kernel_manager.update_env(kernel_id=kernel_id, env=self.get_kernel_env(path, name))
+            except TypeError:
+                self.kernel_manager.update_env(kernel_id="waiting", env=self.get_kernel_env("Untitled.ipynb", "Waiting for kernel"))
 
     async def kernel_culled(self, kernel_id: str) -> bool:
         """Checks if the kernel is still considered alive and returns true if its not found."""
@@ -472,28 +556,36 @@ class SessionManager(LoggingConfigurable):
 
     async def row_to_model(self, row, tolerate_culled=False):
         """Takes sqlite database session row and turns it into a dictionary"""
-        kernel_culled: bool = await ensure_async(self.kernel_culled(row["kernel_id"]))
-        if kernel_culled:
-            # The kernel was culled or died without deleting the session.
-            # We can't use delete_session here because that tries to find
-            # and shut down the kernel - so we'll delete the row directly.
-            #
-            # If caller wishes to tolerate culled kernels, log a warning
-            # and return None.  Otherwise, raise KeyError with a similar
-            # message.
-            self.cursor.execute("DELETE FROM session WHERE session_id=?", (row["session_id"],))
-            msg = (
-                "Kernel '{kernel_id}' appears to have been culled or died unexpectedly, "
-                "invalidating session '{session_id}'. The session has been removed.".format(
-                    kernel_id=row["kernel_id"], session_id=row["session_id"]
-                )
-            )
+        kernel_id = row["kernel_id"]
+        if kernel_id == "waiting":
             if tolerate_culled:
-                self.log.warning(f"{msg}  Continuing...")
                 return None
-            raise KeyError(msg)
+            else:
+                kernel_model = waiting_kernel()
+        else:
+            kernel_culled: bool = await ensure_async(self.kernel_culled(kernel_id))
+            if kernel_culled:
+                # The kernel was culled or died without deleting the session.
+                # We can't use delete_session here because that tries to find
+                # and shut down the kernel - so we'll delete the row directly.
+                #
+                # If caller wishes to tolerate culled kernels, log a warning
+                # and return None.  Otherwise, raise KeyError with a similar
+                # message.
+                self.cursor.execute("DELETE FROM session WHERE session_id=?", (row["session_id"],))
+                msg = (
+                    "Kernel '{kernel_id}' appears to have been culled or died unexpectedly, "
+                    "invalidating session '{session_id}'. The session has been removed.".format(
+                        kernel_id=row["kernel_id"], session_id=row["session_id"]
+                    )
+                )
+                if tolerate_culled:
+                    self.log.warning(f"{msg}  Continuing...")
+                    return None
+                raise KeyError(msg)
 
-        kernel_model = await ensure_async(self.kernel_manager.kernel_model(row["kernel_id"]))
+            kernel_model = await ensure_async(self.kernel_manager.kernel_model(kernel_id))
+
         model = {
             "id": row["session_id"],
             "path": row["path"],
@@ -516,7 +608,8 @@ class SessionManager(LoggingConfigurable):
         for row in c.fetchall():
             try:
                 model = await self.row_to_model(row)
-                result.append(model)
+                if model["kernel"]["id"] != "waiting":
+                    result.append(model)
             except KeyError:
                 pass
         return result
